Inline NoConstructorAgrumentsForInterfaceMock_1
